+++ Code/Source/DioramaModuleInterface.cpp
@@ -14,6 +14,7 @@
 #include <Clients/Collision2DSystemComponent.h>
 #include <Clients/DioramaCamera2DComponent.h>
 #include <Clients/DioramaLightComponent.h>
+#include <Clients/DioramaParallaxComponent.h>
 #include <Clients/ParticleEmitterComponent.h>
 #include <Clients/DioramaSystemComponent.h>
 #include <Clients/SpriteComponent.h>
@@ -42,6 +43,7 @@ namespace Diorama
                 DioramaLightComponent::CreateDescriptor(),
                 DioramaCamera2DComponent::CreateDescriptor(),
                 ParticleEmitterComponent::CreateDescriptor(),
+                DioramaParallaxComponent::CreateDescriptor(),
             });
     }